Rename the cmd package to commands
