+++ commands/k8s/exec.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/docker/cli/cli/streams"
-	"github.com/shipyard/shipyard-cli/cmd/services"
+	"github.com/shipyard/shipyard-cli/commands/services"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	v1 "k8s.io/api/core/v1"
