+++ commands/services/check.go
